TimeInput: Move autocomplete logic to helpers and export it from the library
