+++ src/index.tsx
@@ -156,6 +156,7 @@ export {
   SpacingProp,
 } from './core/theme';
 export { getLogger, setLogger, Logger } from './utils/log/logger';
+export { autocompleteTimeString } from './utils/common';
 export {
   suomifiDesignTokens,
   DesignTokens,
